My First Python Code "The Rando Number Game"
At first I was making a random number generator code but it's kinda get boring real quick so I added some spice to it such as yes/no promt and name promt (also assign specific name as an easter egg), it might not be that much, programming-wise. but it's very rewarding for a beginner like me.